fix(auth): updateEmail
